+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 2.5.1 (April 11, 2022)
+
+FIX:
+* Only run OCI login on create and update (#846)
+* OCI login concurrency issue (#848)
+
+## 2.5.0 (March 28, 2022)
+
+* Upgrade helm dependency to v3.8.1
+* Add support for OCI registries 
+
 ## 2.4.1 (November 09, 2021)
 
 HOTFIX: